wasm2c: Enable exceptions and simd according to command line flags
