+++ src/prebuilt/wasm2c_header_top.cc
@@ -4,6 +4,13 @@ R"w2c_template(
 #include "wasm-rt.h"
 )w2c_template"
 R"w2c_template(
+#if defined(WASM_RT_ENABLE_EXCEPTION_HANDLING)
+)w2c_template"
+R"w2c_template(#include "wasm-rt-exceptions.h"
+)w2c_template"
+R"w2c_template(#endif
+)w2c_template"
+R"w2c_template(
 #if defined(WASM_RT_ENABLE_SIMD)
 )w2c_template"
 R"w2c_template(#include "simde/wasm/simd128.h"
