+++ test/wasm2c/check-imports.txt
@@ -21,9 +21,15 @@
 #ifndef WASM_H_GENERATED_
 #define WASM_H_GENERATED_
 
+#define WASM_RT_ENABLE_SIMD
+
 #include <stdint.h>
 
 #include "wasm-rt.h"
+
+#if defined(WASM_RT_ENABLE_EXCEPTION_HANDLING)
+#include "wasm-rt-exceptions.h"
+#endif
 
 #if defined(WASM_RT_ENABLE_SIMD)
 #include "simde/wasm/simd128.h"
