+++ wasm2c/examples/fac/fac.h
@@ -6,6 +6,10 @@
 
 #include "wasm-rt.h"
 
+#if defined(WASM_RT_ENABLE_EXCEPTION_HANDLING)
+#include "wasm-rt-exceptions.h"
+#endif
+
 #if defined(WASM_RT_ENABLE_SIMD)
 #include "simde/wasm/simd128.h"
 #endif
