+++ wasm2c/wasm-rt.h
@@ -318,6 +318,38 @@ bool wasm_rt_is_initialized(void);
 void wasm_rt_free(void);
 
 /**
+ * A hardened jmp_buf that allows checking for initialization before use
+ */
+typedef struct {
+  /* Is the jmp buf intialized? */
+  bool initialized;
+  /* jmp_buf contents */
+  jmp_buf buffer;
+} wasm_rt_jmp_buf;
+
+#if WASM_RT_INSTALL_SIGNAL_HANDLER && !defined(_WIN32)
+#define WASM_RT_SETJMP_SETBUF(buf) sigsetjmp(buf, 1)
+#else
+#define WASM_RT_SETJMP_SETBUF(buf) setjmp(buf)
+#endif
+
+#define WASM_RT_SETJMP(buf) \
+  ((buf).initialized = true, WASM_RT_SETJMP_SETBUF((buf).buffer))
+
+#if WASM_RT_INSTALL_SIGNAL_HANDLER && !defined(_WIN32)
+#define WASM_RT_LONGJMP_UNCHECKED(buf, val) siglongjmp(buf, val)
+#else
+#define WASM_RT_LONGJMP_UNCHECKED(buf, val) longjmp(buf, val)
+#endif
+
+#define WASM_RT_LONGJMP(buf, val)                                  \
+  /* Abort on failure as this may be called in the trap handler */ \
+  if (!((buf).initialized))                                        \
+    abort();                                                       \
+  (buf).initialized = false;                                       \
+  WASM_RT_LONGJMP_UNCHECKED((buf).buffer, val)
+
+/**
  * Stop execution immediately and jump back to the call to `wasm_rt_impl_try`.
  * The result of `wasm_rt_impl_try` will be the provided trap reason.
  *
@@ -329,72 +361,6 @@ WASM_RT_NO_RETURN void wasm_rt_trap(wasm_rt_trap_t);
  * Return a human readable error string based on a trap type.
  */
 const char* wasm_rt_strerror(wasm_rt_trap_t trap);
-
-/**
- * A tag is represented as an arbitrary pointer.
- */
-typedef const void* wasm_rt_tag_t;
-
-/**
- * Set the active exception to given tag, size, and contents.
- */
-void wasm_rt_load_exception(const wasm_rt_tag_t tag,
-                            uint32_t size,
-                            const void* values);
-
-/**
- * Throw the active exception.
- */
-WASM_RT_NO_RETURN void wasm_rt_throw(void);
-
-/**
- * A hardened jmp_buf that allows us to checks if it is initialized before use
- */
-typedef struct {
-  /* Is the jmp buf intialized? */
-  bool initialized;
-  /* jmp_buf contents */
-  jmp_buf buffer;
-} wasm_rt_jmp_buf;
-
-/**
- * The type of an unwind target if an exception is thrown and caught.
- */
-#define WASM_RT_UNWIND_TARGET wasm_rt_jmp_buf
-
-/**
- * Get the current unwind target if an exception is thrown.
- */
-WASM_RT_UNWIND_TARGET* wasm_rt_get_unwind_target(void);
-
-/**
- * Set the unwind target if an exception is thrown.
- */
-void wasm_rt_set_unwind_target(WASM_RT_UNWIND_TARGET* target);
-
-/**
- * Tag of the active exception.
- */
-wasm_rt_tag_t wasm_rt_exception_tag(void);
-
-/**
- * Size of the active exception.
- */
-uint32_t wasm_rt_exception_size(void);
-
-/**
- * Contents of the active exception.
- */
-void* wasm_rt_exception(void);
-
-#if WASM_RT_INSTALL_SIGNAL_HANDLER && !defined(_WIN32)
-#define WASM_RT_SETJMP_SETBUF(buf) sigsetjmp(buf, 1)
-#else
-#define WASM_RT_SETJMP_SETBUF(buf) setjmp(buf)
-#endif
-
-#define WASM_RT_SETJMP(buf) \
-  ((buf).initialized = true, WASM_RT_SETJMP_SETBUF((buf).buffer))
 
 #define wasm_rt_try(target) WASM_RT_SETJMP(target)
